+++ ChangeLog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version [0.21.1](https://github.com/cedx/base/compare/v0.21.0...v0.21.1)
+- Fixed issues in the `UI.Components.DialogButton` and `UI.Components.DialogBox` components.
+
 ## Version [0.21.0](https://github.com/cedx/base/compare/v0.20.0...v0.21.0)
 - Added the `show()` and `hide()` methods to the `UI.Components.OfflineIndicator` component.
 - Renamed the `UI.Components.MessageBox` component to `DialogBox`.